requirements.txt
requests
python-dotenv
OTXv2